move things around and comment better
