+++ modules/zuora/src/objectQuery/queries/types.ts
@@ -7,8 +7,14 @@ Using these gives type safety and auto complete.
  */
 
 /**
- * One entry in an expand registry: an object mapping the single response field
- * name to its zod schema, e.g. `{ subscriptions: z.array(...) }`.
+ * One entry in an expand registry.
+ *
+ * This should usually be a single-key object where the key is the response
+ * field populated by that expand and the value is the zod schema for that
+ * response field, e.g. `{ subscriptions: z.array(...) }`.
+ *
+ * Duplicate response keys across selected expands are invalid because they
+ * would collide in the parsed response shape.
  */
 export type ExpandRegistryEntry = Record<string, ZodTypeAny>;
 export type ObjectQueryFieldRegistry = Record<string, ZodTypeAny>;
